split node.rs into local and remote cases
